Rest version1

+++ src/main/java/com/example/todoist/controller/SectionController.java
@@ -1,6 +1,7 @@
 package com.example.todoist.controller;
 
 import com.example.todoist.model.Section;
+import com.example.todoist.service.ProjectService;
 import com.example.todoist.service.SectionService;
 import com.example.todoist.requestBean.ServiceRequest;
 import com.example.todoist.responseBean.SectionResponse;
@@ -16,6 +17,18 @@ public class SectionController {
     @Autowired
     SectionService sectionService;
 
+    @Autowired
+    ProjectService projectService;
+
+    boolean checkValidSectionInput(String sectionName,Integer projectId)
+    {
+        if(sectionName==null || sectionName.trim().length()==0 || projectId==null)
+        {
+            return false;
+        }
+        return true;
+    }
+
     @GetMapping("/sections")
     @ResponseBody
     private ResponseEntity getAllSections()
@@ -27,12 +40,27 @@ public class SectionController {
     @ResponseBody
     private ResponseEntity createNewSection(@RequestBody ServiceRequest serviceRequest)
     {
-        Section section;
 
-        if(serviceRequest.getOrder()==null)
-            section=new Section(serviceRequest.getName(),serviceRequest.getProject_id());
-        else
-            section=new Section(serviceRequest.getName(),serviceRequest.getProject_id(),serviceRequest.getOrder());
+        //Check if the Section Input is valid or Not
+        if(!checkValidSectionInput(serviceRequest.getName(),serviceRequest.getProject_id()))
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        }
+
+
+        Section section;
+        String serviceRequestName=serviceRequest.getName().trim();
+
+        if(serviceRequest.getOrder()==null) {
+            section = new Section(serviceRequestName, serviceRequest.getProject_id());
+        }
+        else {
+            Integer id=serviceRequest.getProject_id();
+            if(projectService.findProjectById(id).getId()!=null)
+                section = new Section(serviceRequestName, serviceRequest.getProject_id(), serviceRequest.getOrder());
+            else
+                section = new Section(serviceRequestName, 0, serviceRequest.getOrder());
+        }
         sectionService.saveSection(section);
         SectionResponse sectionResponse=new SectionResponse();
         sectionResponse.setId(section.getId());
@@ -46,28 +74,51 @@ public class SectionController {
     @ResponseBody
     private ResponseEntity getSingleSection(@PathVariable("id")Integer id)
     {
-        return new ResponseEntity(sectionService.getSectionById(id),HttpStatus.OK);
+        if(sectionService.getSectionById(id).getId()!=null)
+            return new ResponseEntity(sectionService.getSectionById(id),HttpStatus.OK);
+        else
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping("/sections/{id}")
     private ResponseEntity updateSection(@PathVariable("id")Integer id,@RequestBody ServiceRequest serviceRequest)
     {
-        Section section=sectionService.getOneSectionById(id);
-        SectionResponse sectionResponse=new SectionResponse();
-        sectionResponse.setId(section.getId());
-        sectionResponse.setProject_id(section.getProjectId());
-        sectionResponse.setOrder(section.getSectionOrder());
-        sectionResponse.setName(section.getName());
-        return new ResponseEntity(HttpStatus.NO_CONTENT);
+
+        if(!checkValidSectionInput(serviceRequest.getName(),serviceRequest.getProject_id()))
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        }
+
+
+        if(sectionService.getSectionById(id).getId()!=null)
+        {
+            Section section = sectionService.getOneSectionById(id);
+            SectionResponse sectionResponse = new SectionResponse();
+            sectionResponse.setId(section.getId());
+            sectionResponse.setProject_id(section.getProjectId());
+            sectionResponse.setOrder(section.getSectionOrder());
+            sectionResponse.setName(section.getName().trim());
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
+        else
+        {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
     }
 
 
     @DeleteMapping("/sections/{id}")
     private ResponseEntity deleteSection(@PathVariable("id")Integer id)
     {
-
-        sectionService.deleteSectionById(id);
-        return new ResponseEntity(HttpStatus.NO_CONTENT);
+        if(sectionService.getSectionById(id).getId()!=null)
+        {
+            sectionService.deleteSectionById(id);
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
+        else
+        {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
 
     }
